Validate and coerce as a DataSchema object with specific format restrictions

Checks if an object is a valid DataSchema with specific format restrictions for compatibility with other Maelstrom Research software and returns it with the appropriate Rmonize::class attribute. This function mainly helps validate inputs within other functions of the package but could be used separately to ensure that an object has an appropriate structure.

as_dataschema_mlstr(object)

Arguments

object

A potential DataSchema object to be coerced.

Value

A list of data frame(s) named 'Variables' and (if any) 'Categories', with Rmonize::class 'dataschema_mlstr'.

Details

A DataSchema is the list of core variables to generate across datasets and related metadata. A DataSchema object is a list of data frames with elements named 'Variables' (required) and 'Categories' (if any). The 'Variables' element must contain at least the name column, and the 'Categories' element must contain at least the variable and name columns to be usable in any function. In 'Variables' the name column must also have unique entries, and in 'Categories' the combination of variable and name columns must also be unique.

The object may be specifically formatted to be compatible with additional Maelstrom Research software, in particular Opal environments.

Examples

{

# Use Rmonize_examples to run examples.
library(dplyr)

dataschema_mlstr <- as_dataschema_mlstr(Rmonize_examples$`DataSchema`)
glimpse(dataschema_mlstr)

}
#> List of 2
#>  $ Variables : tibble [9 × 12] (S3: tbl_df/tbl/data.frame)
#>   ..$ index                     : int [1:9] 1 2 3 4 5 6 7 8 9
#>   ..$ name                      : chr [1:9] "adm_unique_id" "adm_study_id" "sdc_age_m" "sdc_marital_m" ...
#>   ..$ label                     : chr [1:9] "Participant identifier" "Study ID" "Age of mother at first visit" "Marital status of mother at first visit" ...
#>   ..$ valueType                 : chr [1:9] "text" "text" "integer" "integer" ...
#>   ..$ unit                      : chr [1:9] NA NA "years" NA ...
#>   ..$ categorical               : chr [1:9] "0" "1" "0" "1" ...
#>   ..$ target_entity             : chr [1:9] "Mother" "Study" "Mother" "Mother" ...
#>   ..$ target_time_period        : chr [1:9] NA NA "At beginning of pregnancy" "At beginning of pregnancy" ...
#>   ..$ acceptable_source         : chr [1:9] "N/A" "N/A" "Questionnaire or medical charts" "Questionnaire" ...
#>   ..$ acceptable_informant      : chr [1:9] "N/A" "N/A" "Participant or other" "Participant" ...
#>   ..$ acceptable_wording        : chr [1:9] "N/A" "N/A" "No restrictions" "No restrictions" ...
#>   ..$ harmonization_instructions: chr [1:9] NA NA "Use information from mother's first visit during pregnancy." "Use information from mother's first visit during pregnancy." ...
#>  $ Categories: tibble [15 × 4] (S3: tbl_df/tbl/data.frame)
#>   ..$ variable: chr [1:15] "adm_study_id" "adm_study_id" "adm_study_id" "adm_study_id" ...
#>   ..$ name    : chr [1:15] "1" "2" "3" "4" ...
#>   ..$ label   : chr [1:15] "Study 1" "Study 2" "Study 3" "Study 4" ...
#>   ..$ missing : logi [1:15] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>  - attr(*, "madshapR::class")= chr "data_dict_mlstr"
#>  - attr(*, "Rmonize::class")= chr "dataschema_mlstr"

<