Validate and coerce as a harmonized dossier object

Checks if an object is a valid harmonized dossier and returns it with the appropriate Rmonize::class attribute. This function mainly helps validate inputs within other functions of the package but could be used separately to ensure that an object has an appropriate structure. The function has two arguments that can optionally be declared by the user (unique_col_dataset and unique_col_id). unique_col_dataset refers to the columns which contains name of each harmonized dataset. unique_col_id refers to the column in harmonized datasets which identifies unique combinations of observation/dataset. These two columns are added to ensure that there is always a unique entity identifier when datasets are pooled.

as_harmonized_dossier(
  object,
  dataschema = attributes(object)$`Rmonize::DataSchema`,
  data_proc_elem = attributes(object)$`Rmonize::Data Processing Elements`,
  harmonized_col_id = attributes(object)$`Rmonize::harmonized_col_id`,
  harmonized_col_dataset = attributes(object)$`Rmonize::harmonized_col_dataset`,
  harmonized_data_dict_apply = FALSE
)

Arguments

object

A A potential harmonized dossier object to be coerced.

dataschema

A DataSchema object.

data_proc_elem

A Data Processing Elements object.

harmonized_col_id

A character string identifying the name of the column present in every dataset to use as a participant identifier.

harmonized_col_dataset

A character string identifying the column to use for dataset names.

harmonized_data_dict_apply

Whether to apply the dataschema to each harmonized dataset. FALSE by default.

Value

A list of data frame(s) containing harmonized dataset(s). The DataSchema and Data Processing Elements are preserved as attributes of the harmonized dossier.

Details

A harmonized dossier is a named list containing one or more data frames, which are harmonized datasets. A harmonized dossier is generally the product of applying processing to a dossier object The name of each harmonized dataset (data frame) is taken from the reference input dataset. A harmonized dossier also contains the DataSchema and Data Processing Elements used in processing as attributes.

A DataSchema is the list of core variables to generate across datasets and related metadata. A DataSchema object is a list of data frames with elements named 'Variables' (required) and 'Categories' (if any). The 'Variables' element must contain at least the name column, and the 'Categories' element must contain at least the variable and name columns to be usable in any function. In 'Variables' the name column must also have unique entries, and in 'Categories' the combination of variable and name columns must also be unique.

The Data Processing Elements specifies the input elements and processing algorithms to generate harmonized variables in the DataSchema formats. It is also contains metadata used to generate documentation of the processing. A Data Processing Elements object is a data frame with specific columns used in data processing: dataschema_variable, input_dataset, input_variables, Mlstr_harmo::rule_category and Mlstr_harmo::algorithm. To initiate processing, the first entry must be the creation of a harmonized primary identifier variable (e.g., participant unique ID).

Examples

{

# Use Rmonize_examples to run examples.
 
library(dplyr)

harmonized_dossier <- Rmonize_examples[["harmonized_dossier"]]
harmonized_dossier <- as_harmonized_dossier(harmonized_dossier)

glimpse(harmonized_dossier$dataset_study1)

}
#> Rows: 4,000
#> Columns: 9
#> $ adm_unique_id             <chr> "PID7216", "PID5380", "PID5531", "PID7612", …
#> $ adm_study_id              <chr+lbl> "1", "1", "1", "1", "1", "1", "1", "1", …
#> $ sdc_age_m                 <int> 27, 38, 27, 26, 33, 31, 32, 29, 29, 26, 33, …
#> $ sdc_marital_m             <int+lbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1…
#> $ lsb_alc_binge_m_preg      <int+lbl> 0, 0, 0, 0, NA, 0, 0, 0, 0, 0, 0, 0, 0, …
#> $ lsb_alc_binge_m_first_tri <int+lbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, …
#> $ preg_gestational_age_del  <int> 41, 38, 39, 41, 40, 38, 38, 39, 41, 40, 41, …
#> $ sdc_sex_c                 <int+lbl> 1, 2, 1, 1, 1, 2, 1, 1, 1, 1, 1, 2, 1, 1…
#> $ pm_birthweight            <dbl> 2603.164, 3295.346, 3426.888, 3112.095, 2394…