Generate harmonized dataset(s) and associated metadata

Reads a DataSchema and Data Processing Elements to generate a harmonized dossier from input dataset(s) in a dossier and associated metadata. The function has one argument that can optionally be declared by the user (unique_col_dataset). It refers to the columns which contains name of each harmonized dataset. These two columns are added to ensure that there is always a unique entity identifier when datasets are pooled.

harmo_process(
  dossier,
  dataschema = dataschema_extract(data_proc_elem),
  data_proc_elem,
  harmonized_col_dataset = NULL
)

Arguments

dossier

List of data frame(s) containing input dataset(s).

dataschema

A DataSchema object.

data_proc_elem

A Data Processing Elements object.

harmonized_col_dataset

A character string identifying the column to use for dataset names. NULL by default.

Value

A list of data frame(s), containing harmonized dataset(s). The DataSchema and Data Processing Elements are preserved as attributes of the output harmonized dossier.

Details

A dossier is a named list containing one or more data frames, which are input datasets. The name of each data frame in the dossier will be used as the name of the associated harmonized dataset produced by harmo_process().

A DataSchema is the list of core variables to generate across datasets and related metadata. A DataSchema object is a list of data frames with elements named 'Variables' (required) and 'Categories' (if any). The 'Variables' element must contain at least the name column, and the 'Categories' element must contain at least the variable and name columns to be usable in any function. In 'Variables' the name column must also have unique entries, and in 'Categories' the combination of variable and name columns must also be unique.

The Data Processing Elements specifies the algorithms used to process input variables into harmonized variables in the DataSchema format. It is also contains metadata used to generate documentation of the processing. A Data Processing Elements object is a data frame with specific columns used in data processing: dataschema_variable, input_dataset, input_variables, Mlstr_harmo::rule_category and Mlstr_harmo::algorithm. To initiate processing, the first entry must be the creation of a harmonized primary identifier variable (e.g., participant unique ID).

Examples

{

# Use Rmonize_DEMO to run examples.

library(dplyr)
library(madshapR) # data_dict_filter

dataset_MELBOURNE <- Rmonize_DEMO$dataset_MELBOURNE[1]
dossier <- dossier_create(list(dataset_MELBOURNE))

dataschema <- 
  Rmonize_DEMO$`dataschema - final` %>%
  data_dict_filter('name == "adm_unique_id"')

data_proc_elem <- Rmonize_DEMO$`data_processing_elements - final` %>%
  dplyr::filter(dataschema_variable == 'adm_unique_id',
         input_dataset == 'dataset_MELBOURNE')

# perform harmonization
harmo_process(dossier,dataschema,data_proc_elem)

}
#> - Data Processing Elements: ------------------------------------------------
#> 
#> --harmonization of : dataset_MELBOURNE --------------------------------
#>     processing 1/1 : adm_unique_id                id created
#> 
#> 
#> - CREATION OF HARMONIZED DATA DICTIONARY : --------------------------------
#> 
#> dataset_MELBOURNE : done
#> 
#> - WARNING MESSAGES (if any): ----------------------------------------------
#> 
#> $dataset_MELBOURNE
#> # A tibble: 19 × 1
#>    adm_unique_id
#>  * <chr>        
#>  1 377943       
#>  2 497013       
#>  3 927676       
#>  4 995667       
#>  5 21829        
#>  6 209432       
#>  7 272983       
#>  8 580632       
#>  9 304624       
#> 10 637551       
#> 11 279817       
#> 12 235415       
#> 13 373673       
#> 14 485098       
#> 15 299427       
#> 16 854073       
#> 17 197666       
#> 18 130327       
#> 19 220050       
#> 
#> attr(,"Rmonize::class")
#> [1] "harmonized_dossier"
#> attr(,"Rmonize::DataSchema")
#> attr(,"Rmonize::DataSchema")$Variables
#> # A tibble: 1 × 4
#>   name          `label:en`                                     valueType index
#>   <chr>         <chr>                                          <chr>     <chr>
#> 1 adm_unique_id Unique identification code of the participant. text      1    
#> 
#> attr(,"Rmonize::DataSchema")attr(,"madshapR::class")
#> [1] "data_dict_mlstr"
#> attr(,"Rmonize::DataSchema")attr(,"Rmonize::class")
#> [1] "dataschema_mlstr"
#> attr(,"Rmonize::Data Processing Elements")
#> # A tibble: 1 × 11
#>   index dataschema_variable valueType input_dataset     input_variables
#> * <dbl> <chr>               <chr>     <chr>             <chr>          
#> 1     1 adm_unique_id       NA        dataset_MELBOURNE id             
#> # ℹ 6 more variables: `Mlstr_harmo::rule_category` <chr>,
#> #   `Mlstr_harmo::algorithm` <chr>, `Mlstr_harmo::status` <chr>,
#> #   `Mlstr_harmo::status_detail` <chr>, `Mlstr_harmo::comment` <chr>,
#> #   `Rmonize::r_script` <chr>
#> attr(,"Rmonize::harmonized_col_id")
#> [1] "adm_unique_id"