Print a summary of data processing in the console

Reads a harmonized dossier, product of harmo_process(), to list processes, any errors, and an overview of each harmonization rule. The output printed in the console can help in correcting any errors that occurred during data processing.

show_harmo_error(harmonized_dossier, show_warnings = TRUE)

Arguments

harmonized_dossier

A list containing the harmonized dataset(s).

show_warnings

Whether the function should print warnings or not. TRUE by default.

Value

Nothing to be returned. The function prints messages in the console, showing any errors in data processing.

Details

A harmonized dossier is a named list containing one or more data frames, which are harmonized datasets. A harmonized dossier is generally the product of applying processing to a dossier object The name of each harmonized dataset (data frame) is taken from the reference input dataset. A harmonized dossier also contains the DataSchema and Data Processing Elements used in processing as attributes.

Examples

{
# Use Rmonize_examples to run examples.
library(dplyr)

# Perform data processing
harmonized_dossier <- Rmonize_examples$`harmonized_dossier`

# Show error(s) on the console
show_harmo_error(harmonized_dossier)
}
#> 
#> 
#> - STATUS SUMMARY: ----------------------------------------------------
#> 
#> # A tibble: 8 × 3
#>   `Mlstr_harmo::rule_category` `Total number of algorithms` success
#>   <chr>                        <chr>                        <chr>  
#> 1 case_when                    6                            100 %  
#> 2 direct_mapping               7                            100 %  
#> 3 id_creation                  5                            100 %  
#> 4 impossible                   5                            100 %  
#> 5 operation                    6                            100 %  
#> 6 other                        1                            100 %  
#> 7 paste                        5                            100 %  
#> 8 recode                       10                           100 %  
#>