Generate harmonized dataset(s) and annotated Data Processing Elements

Reads the DataSchema and Data Processing Elements objects to generate harmonized dataset(s) and annotated Data Processing Elements object with harmonization statuses and any processing errors. The function uses the DataSchema and Data Processing Elements specifications to process input variables into output harmonized variables for each dataset. Documentation of each data processing action is generated in the console to support the identification of errors and correction of the Data Processing Elements file and objects, as needed. An annotated Data Processing Elements is also produced, providing harmonization statuses (complete/impossible) for each DataSchema variable in input dataset, which can be used to create a summary of the harmonization potential of the DataSchema variables across input dataset(s).

harmo_process(dossier, dataschema = NULL, data_proc_elem)

Arguments

dossier

List of tibble(s), each of them being datasets to be harmonised.

dataschema

A list of tibble(s) representing metadata of an associated harmonized dossier.

data_proc_elem

A tibble, identifying the input Data Processing Elements.

Value

A list of tibbles of each harmonized dataset that has been harmonized from input datataset.

Details

A dossier must be a named list containing at least one data frame or data frame extension (e.g. a tibble), each of them being datasets. The name of each tibble will be use as the reference name of the dataset.

A DataSchema defines the harmonized variables to be generated, and also represents metadata of an associated harmonized dossier. It must be a list of data frame or data frame extension (e.g. a tibble) objects with elements named "Variables" (required) and "Categories" (if any). The "Variables" element must contain at least the name column, and the "Categories" element must contain at least the variable and name columns to be usable in any function. To be considered as a minimum workable DataSchema, in "Variables" the name column must also have unique and non-null entries, and in "Categories" the combination of variable and name columns must also be unique.

A Data Processing Elements contains the rules and metadata that will be used to perform harmonization of input datasets in accordance with the DataSchema. It must be a data frame or data frame extension (e.g. a tibble) and it must contain certain columns which participate to the process, including the dataschema_variable, input_dataset,input_variables, Mlstr_harmo::rule_category and Mlstr_harmo::algorithm. The mandatory first processing element must be "id_creation" in Mlstr_harmo::rule_category followed by the name of the column taken as identifier of each dataset to initiate the process of harmonization.

Examples

{

# You can use our demonstration files to run examples

library(dplyr)
library(madshapR) # data_dict_filter

dataset_MELBOURNE_1 <- DEMO_files_harmo$dataset_MELBOURNE_1[1]
dossier <- dossier_create(list(dataset_MELBOURNE_1))

dataschema <- 
  DEMO_files_harmo$`dataschema - final` %>%
  data_dict_filter('name == "adm_unique_id"')

data_proc_elem <- DEMO_files_harmo$`data_processing_elements - final` %>%
  dplyr::filter(dataschema_variable == 'adm_unique_id',
         input_dataset == 'dataset_MELBOURNE_1')

# perform harmonization
harmo_process(dossier,dataschema,data_proc_elem)

}
#> - Data Processing Elements: ------------------------------------------------
#> 
#> --harmonization of : dataset_MELBOURNE_1 ------------------------------
#>     processing 1/1 : adm_unique_id                id created
#> 
#> 
#> - CREATION OF HARMONIZED DATA DICTIONARY : --------------------------------
#> 
#> dataset_MELBOURNE_1 : done
#> 
#> 
#> ------------------------------------------------------------------------------
#> 
#> Your harmonization is done. Please check if everything worked correctly.
#> 
#> - WARNING MESSAGES (if any): ----------------------------------------------
#> 
#> $dataset_MELBOURNE_1
#> # A tibble: 19 × 1
#>    adm_unique_id
#>    <chr>        
#>  1 377943       
#>  2 497013       
#>  3 927676       
#>  4 995667       
#>  5 21829        
#>  6 209432       
#>  7 272983       
#>  8 580632       
#>  9 304624       
#> 10 637551       
#> 11 279817       
#> 12 235415       
#> 13 373673       
#> 14 485098       
#> 15 299427       
#> 16 854073       
#> 17 197666       
#> 18 130327       
#> 19 220050       
#> 
#> attr(,"Rmonize::class")
#> [1] "harmonized_dossier"
#> attr(,"Rmonize::DataSchema")
#> attr(,"Rmonize::DataSchema")$Variables
#> # A tibble: 1 × 6
#>   name          `label:en`  valueType index `Mlstr_area::1` `Mlstr_area::1.term`
#>   <chr>         <chr>       <chr>     <chr> <chr>           <chr>               
#> 1 adm_unique_id Unique ide… text      1     ADM             Identifiers         
#> 
#> attr(,"Rmonize::DataSchema")attr(,"madshapR::class")
#> [1] "data_dict_mlstr"
#> attr(,"Rmonize::DataSchema")attr(,"Rmonize::class")
#> [1] "dataSchema_mlstr"
#> attr(,"Rmonize::Data Processing Elements")
#> # A tibble: 1 × 11
#>   index dataschema_variable valueType input_dataset       input_variables
#> * <dbl> <chr>               <chr>     <chr>               <chr>          
#> 1     1 adm_unique_id       text      dataset_MELBOURNE_1 id             
#> # ℹ 6 more variables: `Mlstr_harmo::rule_category` <chr>,
#> #   `Mlstr_harmo::algorithm` <chr>, `Mlstr_harmo::status` <chr>,
#> #   `Mlstr_harmo::status_detail` <chr>, `Mlstr_harmo::comment` <chr>,
#> #   `Rmonize::r_script` <chr>
#> attr(,"Rmonize::harmonized_col_id")
#> [1] "adm_unique_id"