Generate a web-based bookdown visual report of a harmonized dossier

Generates a visual report for a dataset in an HTML bookdown document. The report provides figures and descriptive statistics for each variable to facilitate the assessment of input data. Statistics and figures are generated according to variable data type. The report can be used to help assess data structure, coherence across elements, and taxonomy or data dictionary formats. The summaries and figures provide additional information about variable distributions and descriptive statistics. The charts and tables are produced based on their data type. The variable can be grouped using group_by parameter, which is a (categorical) column in the dataset. The user may need to use as.factor() in this context. To fasten the process (and allow recycling object in a workflow) the user can feed the function with a .summary_var, which is the output of the function dataset_summarize() of the column(s) col and group_by. The summary must have the same parameters to operate.

harmonized_dossier_visualize(
  harmonized_dossier = NULL,
  pooled_harmonized_dataset = pooled_harmonized_dataset_create(harmonized_dossier),
  dataschema = data_dict_extract(pooled_harmonized_dataset),
  group_by = attributes(pooled_harmonized_dataset)$`Rmonize::unique_col_dataset`,
  bookdown_path,
  taxonomy = NULL,
  valueType_guess = FALSE,
  render = "html",
  .summary_pool = NULL
)

Arguments

harmonized_dossier

List of tibble(s), each of them being harmonized dataset.

pooled_harmonized_dataset

A tibble, identifying the pooled harmonized dataset.

dataschema

A list of tibble(s) representing metadata of an associated harmonized dossier.

group_by

A character string of one column in the pooled harmonized dataset that can be taken as a grouping column. The visual element will be grouped and displayed by this column.

bookdown_path

A character string identifying the folder path where the bookdown report will be saved.

taxonomy

A tibble identifying the scheme used for variables classification.

valueType_guess

Whether the output should include a more accurate valueType that could be applied to the dataset. FALSE by default.

render

Output format of the visual report. To date, the format can only be 'html', but will be expand to other formats in the future.

.summary_pool

A list which is the summary of the variables.

Value

A bookdown folder containing files in the specified output folder. To open the file in browser, open 'docs/index.html'. Or use bookdown_open()

Details

A harmonized dossier must be a named list containing at least one data frame or data frame extension (e.g. a tibble), each of them being harmonized dataset(s). It is generally the product of applying harmonization processing to a dossier object. The name of each tibble will be use as the reference name of the dataset. A harmonized dossier has four attributes : Rmonize::class which is "harmonized_dossier" ; Rmonize::DataSchema (provided by user) ; Rmonize::Data Processing Elements ; Rmonize::harmonized_col_id (provided by user) which refers to the column in each dataset which identifies unique combination observation/dataset. This id column name is the same across the dataset(s), the DataSchema and the Data Processing Elements (created by using 'id_creation') and is used to initiate the process of harmonization.

A taxonomy is classification scheme that can be defined for variable attributes. If defined, a taxonomy must be a data frame like object. It must be compatible with (and is generally extracted from) an Opal environment. To work with certain functions, a valid taxonomy must contain at least the columns 'taxonomy', 'vocabulary', and 'terms'. In addition, the taxonomy may follow Maelstrom research taxonomy, and its content can be evaluated accordingly, such as naming convention restriction, tagging elements, or scales, which are specific to Maelstrom Research. In this particular case, the tibble must also contain 'vocabulary_short', 'taxonomy_scale', 'vocabulary_scale' and 'term_scale' to work with some specific functions.

The valueType is a property of a variable and is required in certain functions to determine the handling of the variables. The valueType refers to the OBiBa-internal type of a variable. It is specified in a data dictionary in a column valueType and can be associated with variables as attributes. Acceptable valueTypes include 'text', 'integer', 'decimal', 'boolean', datetime', 'date'). The full list of OBiBa valueType possibilities and their correspondence with R data types are available using madshapR::valueType_list.

See also

Examples

{

# You can use our demonstration files to run examples

library(madshapR)
library(dplyr)
library(fs)

pooled_harmonized_dataset <- Rmonize_DEMO$pooled_harmonized_dataset
dataschema <- Rmonize_DEMO$`dataschema - final` %>%
  data_dict_filter('name == "adm_unique_id"')

summary_var_harmo <- Rmonize_DEMO$summary_var_harmo

if(dir_exists(tempdir())) dir_delete(tempdir())
bookdown_path <- tempdir()

harmonized_dossier_visualize(
  pooled_harmonized_dataset = pooled_harmonized_dataset,
  dataschema = dataschema,
  bookdown_path = bookdown_path,
  .summary_pool = summary_var_harmo)

# To open the file in browser, open 'bookdown_path/docs/index.html'.
# Or use bookdown_open(bookdown_path) function

}
#> 1/7                  
#> 2/7 [unnamed-chunk-1]
#> 3/7                  
#> 4/7 [unnamed-chunk-2]
#> 5/7                  
#> 6/7 [unnamed-chunk-3]
#> 7/7                  
#> "C:/Program Files/RStudio/resources/app/bin/quarto/bin/tools/pandoc" +RTS -K512m -RTS bookdownproj.knit.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash --output bookdownproj.html --lua-filter "C:\Users\guill\AppData\Local\R\win-library\4.3\bookdown\rmarkdown\lua\custom-environment.lua" --lua-filter "C:\Users\guill\AppData\Local\R\win-library\4.3\rmarkdown\rmarkdown\lua\pagebreak.lua" --lua-filter "C:\Users\guill\AppData\Local\R\win-library\4.3\rmarkdown\rmarkdown\lua\latex-div.lua" --lua-filter "C:\Users\guill\AppData\Local\R\win-library\4.3\rmarkdown\rmarkdown\lua\anchor-sections.lua" --metadata-file "C:\Users\guill\AppData\Local\Temp\Rtmpu2MlPL\file3898316e2ebc" --wrap preserve --standalone --section-divs --table-of-contents --toc-depth 3 --template "C:\Users\guill\AppData\Local\R\win-library\4.3\bookdown\templates\gitbook.html" --highlight-style pygments --number-sections --css style.css --mathjax --include-in-header "C:\Users\guill\AppData\Local\Temp\Rtmpu2MlPL\rmarkdown-str38981e25bb1.html" 
#> 
#> 
#> To edit your file, You can use the function `bookdown_open('C:\Users\guill\AppData\Local\Temp\Rtmpu2MlPL')`
#> (Compatibility tested on Chrome, Edge and Mozilla)
#>